LayoutTests/accessibility/input-search-cancel-button-expected.txt

This tests that there exists an accessibility element for the search cancel button.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.accessibleElementById("search").childAtIndex(1).description is "AXDescription: cancel"
PASS accessibilityController.accessibleElementById("search").childAtIndex(1).role is "AXRole: AXButton"
PASS successfullyParsed is true

TEST COMPLETE
